moved param tests to manual
